+++ Runtime/CreateAssetMenuConstants.cs
@@ -6,7 +6,13 @@ namespace Zor.RandomGenerators
 	{
 		public const string RandomGeneratorProvidersFolder = "Random Generator Providers/";
 
-		public const string ContinuousDistributionsFolder = RandomGeneratorProvidersFolder + "Continuous Distributions/";
+		public const string ContinuousDistributionsFolder =
+			RandomGeneratorProvidersFolder + "Continuous Distributions/";
+
+		public const string IndependentDistributionsFolder =
+			ContinuousDistributionsFolder + "Independent Distributions/";
+		public const string AcceptanceRejectionDistributionFolder =
+			IndependentDistributionsFolder + "Acceptance-Rejection Distribution/";
 
 		public const string NormalDistributionsFolder = ContinuousDistributionsFolder + "Normal Distributions/";
 		public const string BatesDistributionFolder = NormalDistributionsFolder + "Bates Distribution/";